factor into csvReader and csvUtils files
